+++ CHANGELOG.md
@@ -39,6 +39,8 @@ also non string parameters (e.g. `Enum.join([1, 2], ",")`
 - ESP32: add support for `esp_adc` ADC driver, with Erlang and Elixir examples
 - Add handler for ESP32 network driver STA mode `beacon_timeout` (event: 21), see issue
 [#1100](https://github.com/atomvm/AtomVM/issues/1100)
+- Support for mounting/unmounting storage on ESP32 (such as SD or internal flash) using
+`esp:mount/4` and `esp:umount/1`
 
 ### Changed